Patching a strange bug
when an item has no display, but is still purchasable, it seems to cause issues with saving. As such, items without a display are now considered out of stock. I don't see how this will break anything, as purchasable spots always restock eventually on their own anyway.

also adding a public method for when a cock is cumming, because it's kinda just useful (I think)